requirements.txt
aniso8601==9.0.1
click==8.0.3
Flask==2.0.2
Flask-MySQL==1.5.2
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.9
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.27
mysql-connector-python-rf==2.2.2
mysqlclient==2.1.0
protobuf==3.19.1
PyMySQL==1.0.2
pytz==2021.3
six==1.16.0
Werkzeug==2.0.2
